requirements.txt
streamlit
pandas
scholarly
python-docx
openpyxl